+++ src/components/alert/alert-transitions.ts
@@ -7,9 +7,9 @@ import { Transition } from '../../transitions/transition';
  */
 export class AlertPopIn extends Transition {
   init() {
-    let ele = this.enteringView.pageRef().nativeElement;
-    let backdrop = new Animation(this.plt, ele.querySelector('ion-backdrop'));
-    let wrapper = new Animation(this.plt, ele.querySelector('.alert-wrapper'));
+    const ele = this.enteringView.pageRef().nativeElement;
+    const backdrop = new Animation(this.plt, ele.querySelector('ion-backdrop'));
+    const wrapper = new Animation(this.plt, ele.querySelector('.alert-wrapper'));
 
     wrapper.fromTo('opacity', 0.01, 1).fromTo('scale', 1.1, 1);
     backdrop.fromTo('opacity', 0.01, 0.3);
@@ -25,9 +25,9 @@ export class AlertPopIn extends Transition {
 
 export class AlertPopOut extends Transition {
   init() {
-    let ele = this.leavingView.pageRef().nativeElement;
-    let backdrop = new Animation(this.plt, ele.querySelector('ion-backdrop'));
-    let wrapper = new Animation(this.plt, ele.querySelector('.alert-wrapper'));
+    const ele = this.leavingView.pageRef().nativeElement;
+    const backdrop = new Animation(this.plt, ele.querySelector('ion-backdrop'));
+    const wrapper = new Animation(this.plt, ele.querySelector('.alert-wrapper'));
 
     wrapper.fromTo('opacity', 0.99, 0).fromTo('scale', 1, 0.9);
     backdrop.fromTo('opacity', 0.3, 0);
@@ -43,9 +43,9 @@ export class AlertPopOut extends Transition {
 
 export class AlertMdPopIn extends Transition {
   init() {
-    let ele = this.enteringView.pageRef().nativeElement;
-    let backdrop = new Animation(this.plt, ele.querySelector('ion-backdrop'));
-    let wrapper = new Animation(this.plt, ele.querySelector('.alert-wrapper'));
+    const ele = this.enteringView.pageRef().nativeElement;
+    const backdrop = new Animation(this.plt, ele.querySelector('ion-backdrop'));
+    const wrapper = new Animation(this.plt, ele.querySelector('.alert-wrapper'));
 
     wrapper.fromTo('opacity', 0.01, 1).fromTo('scale', 1.1, 1);
     backdrop.fromTo('opacity', 0.01, 0.5);
@@ -61,9 +61,9 @@ export class AlertMdPopIn extends Transition {
 
 export class AlertMdPopOut extends Transition {
   init() {
-    let ele = this.leavingView.pageRef().nativeElement;
-    let backdrop = new Animation(this.plt, ele.querySelector('ion-backdrop'));
-    let wrapper = new Animation(this.plt, ele.querySelector('.alert-wrapper'));
+    const ele = this.leavingView.pageRef().nativeElement;
+    const backdrop = new Animation(this.plt, ele.querySelector('ion-backdrop'));
+    const wrapper = new Animation(this.plt, ele.querySelector('.alert-wrapper'));
 
     wrapper.fromTo('opacity', 0.99, 0).fromTo('scale', 1, 0.9);
     backdrop.fromTo('opacity', 0.5, 0);
@@ -79,9 +79,9 @@ export class AlertMdPopOut extends Transition {
 
 export class AlertWpPopIn extends Transition {
   init() {
-    let ele = this.enteringView.pageRef().nativeElement;
-    let backdrop = new Animation(this.plt, ele.querySelector('ion-backdrop'));
-    let wrapper = new Animation(this.plt, ele.querySelector('.alert-wrapper'));
+    const ele = this.enteringView.pageRef().nativeElement;
+    const backdrop = new Animation(this.plt, ele.querySelector('ion-backdrop'));
+    const wrapper = new Animation(this.plt, ele.querySelector('.alert-wrapper'));
 
     wrapper.fromTo('opacity', 0.01, 1).fromTo('scale', 1.3, 1);
     backdrop.fromTo('opacity', 0.01, 0.5);
@@ -97,9 +97,9 @@ export class AlertWpPopIn extends Transition {
 
 export class AlertWpPopOut extends Transition {
   init() {
-    let ele = this.leavingView.pageRef().nativeElement;
-    let backdrop = new Animation(this.plt, ele.querySelector('ion-backdrop'));
-    let wrapper = new Animation(this.plt, ele.querySelector('.alert-wrapper'));
+    const ele = this.leavingView.pageRef().nativeElement;
+    const backdrop = new Animation(this.plt, ele.querySelector('ion-backdrop'));
+    const wrapper = new Animation(this.plt, ele.querySelector('.alert-wrapper'));
 
     wrapper.fromTo('opacity', 0.99, 0).fromTo('scale', 1, 1.3);
     backdrop.fromTo('opacity', 0.5, 0);