+++ src/components/slides/slides.ts
@@ -65,6 +65,7 @@ import { ViewController } from '../../navigation/view-controller';
  *
  * ```ts
  * import { ViewChild } from '@angular/core';
+ * import { Slides } from 'ionic-angular';
  *
  * class MyPage {
  *   @ViewChild(Slides) slides: Slides;
@@ -104,6 +105,9 @@ import { ViewController } from '../../navigation/view-controller';
  * should be used, where `freeMode` is the option to change:
  *
  * ```ts
+ * import { ViewChild } from '@angular/core';
+ * import { Slides } from 'ionic-angular';
+
  * class MyPage {
  *   @ViewChild(Slides) slides: Slides;
  *
@@ -886,7 +890,14 @@ export class Slides extends Ion {
 
 
 
-  constructor(config: Config, private _plt: Platform, zone: NgZone, @Optional() viewCtrl: ViewController, elementRef: ElementRef, renderer: Renderer) {
+  constructor(
+    config: Config,
+    private _plt: Platform,
+    zone: NgZone,
+    @Optional() viewCtrl: ViewController,
+    elementRef: ElementRef,
+    renderer: Renderer,
+  ) {
     super(config, elementRef, renderer, 'slides');
 
     this._zone = zone;
@@ -960,12 +971,18 @@ export class Slides extends Ion {
     }
   }
 
+  resize() {
+    if (this._init) {
+
+    }
+  }
+
   /**
    * Transition to the specified slide.
    *
    * @param {number} index  The index number of the slide.
    * @param {number} [speed]  Transition duration (in ms).
-   * @param {boolean} [runCallbacks] Whether or not to emit the `ionWillChange`/`ionDidChange` events. Default true.
+   * @param {boolean} [runCallbacks] Whether or not to emit the `ionSlideWillChange`/`ionSlideDidChange` events. Default true.
    */
   slideTo(index: number, speed?: number, runCallbacks?: boolean) {
     slideTo(this, this._plt, index, speed, runCallbacks);
@@ -975,7 +992,7 @@ export class Slides extends Ion {
    * Transition to the next slide.
    *
    * @param {number} [speed]  Transition duration (in ms).
-   * @param {boolean} [runCallbacks]  Whether or not to emit the `ionWillChange`/`ionDidChange` events. Default true.
+   * @param {boolean} [runCallbacks]  Whether or not to emit the `ionSlideWillChange`/`ionSlideDidChange` events. Default true.
    */
   slideNext(speed?: number, runCallbacks?: boolean) {
     slideNext(this, this._plt, runCallbacks, speed, true);
@@ -985,7 +1002,7 @@ export class Slides extends Ion {
    * Transition to the previous slide.
    *
    * @param {number} [speed]  Transition duration (in ms).
-   * @param {boolean} [runCallbacks]  Whether or not to emit the `ionWillChange`/`ionDidChange` events. Default true.
+   * @param {boolean} [runCallbacks]  Whether or not to emit the `ionSlideWillChange`/`ionSlideDidChange` events. Default true.
    */
   slidePrev(speed?: number, runCallbacks?: boolean) {
     slidePrev(this, this._plt, runCallbacks, speed, true);